requirements.txt
pandas
numpy
tweet-preprocessor
ftfy
clean-text[gpl]
tqdm
scikit-learn
scipy
jupyter
gensim
nltk
torch
matplotlib